Remove `'xior/axios'` entry
