+++ tests/src/tests/index.test.ts
@@ -2,35 +2,16 @@ import assert from 'node:assert';
 import { readFile, writeFile } from 'node:fs/promises';
 import { before, after, describe, it } from 'node:test';
 import { stringify } from 'qs';
-import { isXiorError, merge, Xior } from 'xior';
+import { isCancel, isAxiosError as isXiorError, merge, Axios as Xior } from 'xior';
+
 import errorRetryPlugin from 'xior/plugins/error-retry';
 
 import { readChunks, startServer } from './server';
-
-/*
-
-export interface AxiosStatic extends AxiosInstance {
-  Cancel: CancelStatic;
-  CancelToken: CancelTokenStatic;
-  Axios: typeof Axios;
-  AxiosError: typeof AxiosError;
-  HttpStatusCode: typeof HttpStatusCode;
-  readonly VERSION: string;
-  isCancel: typeof isCancel;
-  all: typeof all;
-  spread: typeof spread;
-  isAxiosError: typeof isAxiosError;
-  toFormData: typeof toFormData;
-  formToJSON: typeof formToJSON;
-  getAdapter: typeof getAdapter;
-  CanceledError: typeof CanceledError;
-  AxiosHeaders: typeof AxiosHeaders;
-  mergeConfig: typeof mergeConfig;
-}
-*/
+import { Readable } from 'node:stream';
+import { createWriteStream } from 'node:fs';
 
 let close: Function;
-const port = 7866;
+const port = 7876;
 const baseURL = `http://localhost:${port}`;
 before(async () => {
   close = await startServer(port);
@@ -425,29 +406,25 @@ describe('xior tests', () => {
   describe('stream test', () => {
     it('stream with `responseType: stream` should work', async () => {
       const xiorInstance = Xior.create({ baseURL });
-      const { response } = await xiorInstance.post<{ file: any; body: Record<string, string> }>(
-        '/stream/10',
-        null,
-        { responseType: 'stream' }
-      );
-      const reader = response.body!.getReader();
-      let chunk;
-      for await (chunk of readChunks(reader)) {
-        console.log(`received chunk of size ${chunk.length}`);
-      }
-      assert.strictEqual(chunk.length > 0, true);
+      const { data } = await xiorInstance.post<Readable>('/stream/10', null, {
+        responseType: 'stream',
+      });
+      assert.strictEqual(data.pipe !== undefined, true);
     });
 
     it('stream download image should work', async () => {
       const xiorInstance = Xior.create({ baseURL });
       // GET request for remote image in node.js
       await xiorInstance
-        .get('https://bit.ly/2mTM3nY', {
+        .get<Readable>('https://bit.ly/2mTM3nY', {
           responseType: 'stream',
         })
-        .then(async function ({ response, config }) {
-          const buffer = Buffer.from(await response.arrayBuffer());
-          return writeFile('./uploads/ada_lovelace.jpg', buffer);
+        .then(async function ({ response, config, data }) {
+          const writeStream = createWriteStream('./uploads/ada_lovelace.jpg');
+          data.pipe(writeStream);
+          writeStream.on('finish', () => {
+            console.log('File written successfully');
+          });
         });
 
       assert.strictEqual(1 > 0, true);
@@ -550,6 +527,21 @@ describe('xior tests', () => {
       assert.strictEqual(timeoutError, true);
     });
 
+    it('should work with isCancel when abort error occurs during GET', async () => {
+      const xiorInstance = Xior.create({ baseURL });
+      const controller = new AbortController();
+      setTimeout(() => {
+        controller.abort();
+      }, 1200);
+      let isCancelError = false;
+      try {
+        await xiorInstance.get('/timeout', { signal: controller.signal, timeout: 1000 });
+      } catch (e) {
+        isCancelError = isCancel(e);
+      }
+      assert.strictEqual(isCancelError, true);
+    });
+
     it('Check real timeout', async () => {
       const xiorInstance = Xior.create({ baseURL });
       let timeoutError = false;
@@ -641,7 +633,7 @@ describe('xior tests', () => {
         {},
         { responseType: 'stream' }
       );
-      assert.strictEqual(postData === undefined, true);
+      assert.strictEqual(postData !== undefined, true);
     });
   });
 
@@ -715,7 +707,7 @@ describe('xior tests', () => {
     });
 
     it('Xior.plugins.use/eject/clear should work', () => {
-      assert.strictEqual((xiorInstance as any).P.length, 0);
+      assert.strictEqual((xiorInstance as any).P.length, 1);
       const handler = xiorInstance.plugins.use((plugin) => {
         return (config) => {
           return plugin(config);
@@ -726,9 +718,9 @@ describe('xior tests', () => {
           return plugin(config);
         };
       });
-      assert.strictEqual((xiorInstance as any).P.length, 2);
+      assert.strictEqual((xiorInstance as any).P.length, 3);
       xiorInstance.plugins.eject(handler);
-      assert.strictEqual((xiorInstance as any).P.length, 1);
+      assert.strictEqual((xiorInstance as any).P.length, 2);
 
       xiorInstance.plugins.clear();
       assert.strictEqual((xiorInstance as any).P.length, 0);
@@ -738,7 +730,7 @@ describe('xior tests', () => {
   describe('custom plugins tests', () => {
     const xiorInstance = Xior.create({ baseURL });
     it('xior plugins order should work', async () => {
-      assert.strictEqual((xiorInstance as any).P.length, 0);
+      assert.strictEqual((xiorInstance as any).P.length, 1);
 
       const order: number[] = [];
       xiorInstance.plugins.use((adapter) => {
@@ -761,7 +753,7 @@ describe('xior tests', () => {
           return adapter(config);
         };
       });
-      assert.strictEqual((xiorInstance as any).P.length, 3);
+      assert.strictEqual((xiorInstance as any).P.length, 4);
       await xiorInstance.get('/get');
       assert.strictEqual(order.length, 3);
       assert.strictEqual(order.join(','), [3, 2, 1].join(','));
