+++ tsup.config.ts
@@ -3,14 +3,12 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: {
     'dist/index': 'src/index.ts',
+    'dist/index-node': 'src/index-node.ts',
     utils: 'src/utils.ts',
-    axios: 'src/axios.ts',
-    'node-axios': 'src/node-axios.ts',
     'plugins/error-retry': 'src/plugins/error-retry.ts',
     'plugins/throttle': 'src/plugins/throttle.ts',
     'plugins/dedupe': 'src/plugins/dedupe.ts',
     'plugins/cache': 'src/plugins/cache.ts',
-    'plugins/stream': 'src/plugins/stream.ts',
     'plugins/progress': 'src/plugins/progress.ts',
     'plugins/mock': 'src/plugins/mock.ts',
     'plugins/error-cache': 'src/plugins/error-cache.ts',
